archive/requirements.txt
aiogram
supabase
requests
apscheduler
staticmap
python-dotenv
staticmap==0.5.6
requests==2.31.0
aiohttp==3.9.3
python-dotenv
